CardクラスとDeckクラスを作成
-
Cradクラスだけでなく、Deckクラスも作成し、それぞれが持つべき情報やメソッドを分けた
